# Conflicts:
#	LiXiaoMai/数据库/create.sql

+++ LiXiaoMai/src/main/java/com/example/lixiaomai/backend/entity/Wallet.java
@@ -9,6 +9,6 @@ public class Wallet {
     private String password;
     private double balance;
     private List<Integer> dId;
-    private List<Double> discountNum;
+    private List<Integer> discountNum;
 
 }
